Extract MappingListSerializer
